Added button responsiveness + colors on modal
